+++ api/v1alpha1/cassandra_types.go
@@ -19,7 +19,7 @@ type CassandraSpec struct {
 	DiskSpace string `json:"disk_space,omitempty"`
 
 	// Authentication reference to Aiven token in a secret
-	AuthSecretRef AuthSecretReference `json:"authSecretRef,omitempty"`
+	AuthSecretRef *AuthSecretReference `json:"authSecretRef,omitempty"`
 
 	// Information regarding secret creation
 	ConnInfoSecretTarget ConnInfoSecretTarget `json:"connInfoSecretTarget,omitempty"`
@@ -43,7 +43,7 @@ type Cassandra struct {
 	Status ServiceStatus `json:"status,omitempty"`
 }
 
-func (in *Cassandra) AuthSecretRef() AuthSecretReference {
+func (in *Cassandra) AuthSecretRef() *AuthSecretReference {
 	return in.Spec.AuthSecretRef
 }