+++ api/v1alpha1/kafka_types.go
@@ -17,7 +17,7 @@ type KafkaSpec struct {
 	DiskSpace string `json:"disk_space,omitempty"`
 
 	// Authentication reference to Aiven token in a secret
-	AuthSecretRef AuthSecretReference `json:"authSecretRef,omitempty"`
+	AuthSecretRef *AuthSecretReference `json:"authSecretRef,omitempty"`
 
 	// Information regarding secret creation
 	ConnInfoSecretTarget ConnInfoSecretTarget `json:"connInfoSecretTarget,omitempty"`
@@ -45,7 +45,7 @@ type Kafka struct {
 	Status ServiceStatus `json:"status,omitempty"`
 }
 
-func (in *Kafka) AuthSecretRef() AuthSecretReference {
+func (in *Kafka) AuthSecretRef() *AuthSecretReference {
 	return in.Spec.AuthSecretRef
 }