+++ controllers/kafkatopic_controller_test.go
@@ -114,7 +114,7 @@ func kafkaTopicSpec(service, topic, namespace string) *v1alpha1.KafkaTopic {
 					Value: "val1",
 				},
 			},
-			AuthSecretRef: v1alpha1.AuthSecretReference{
+			AuthSecretRef: &v1alpha1.AuthSecretReference{
 				Name: secretRefName,
 				Key:  secretRefKey,
 			},