+++ api/v1alpha1/postgresql_webhook.go
@@ -59,7 +59,7 @@ func (r *PostgreSQL) ValidateUpdate(old runtime.Object) error {
 func (r *PostgreSQL) ValidateDelete() error {
 	pglog.Info("validate delete", "name", r.Name)
 
-	if r.Spec.TerminationProtection {
+	if r.Spec.TerminationProtection != nil && *r.Spec.TerminationProtection {
 		return errors.New("cannot delete PostgreSQL service, termination protection is on")
 	}