+++ controllers/common.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"context"
+	"errors"
 	"strconv"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -22,7 +23,10 @@ const (
 	instanceIsRunningAnnotation   = "controllers.aiven.io/instance-is-running"
 )
 
-var operatorUserAgent = "k8s-operator/" + aiven.Version()
+var (
+	operatorUserAgent          = "k8s-operator/" + aiven.Version()
+	errTerminationProtectionOn = errors.New("cannot delete object, termination protection is on")
+)
 
 func checkServiceIsRunning(c *aiven.Client, project, serviceName string) (bool, error) {
 	s, err := c.Services.Get(project, serviceName)