+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add new service integrations: `clickhouse_postgresql`, `clickhouse_kafka`, `clickhouse_kafka`, `logs`, `external_aws_cloudwatch_metrics`
 - Add `KafkaTopic.Spec.topicName` field. Unlike the `metadata.name`, supports additional characters and has a longer length.
   `KafkaTopic.Spec.topicName` replaces `metadata.name` in future releases and will be marked as required.
+- Accept `false` value for `termination_protection` property
 
 ## v0.8.0 - 2023-02-15